requirements.txt
python-telegram-bot==20.7
openai==1.3.5
requests>=2.28.0
